refactor: rename NavBarDesktop to NavBar delete useless NavBarItem, use NavMenuItem instead
