+++ src/components/atoms/index.ts
@@ -1,4 +1,5 @@
 export * from "./Button/Button";
+export * from "./ButtonBurger/ButtonBurger";
 export * from "./Flex/Flex";
 export * from "./Logo/Logo";
-export * from "./NavBarItem/NavBarItem";
+export * from "./NavMenuItem/NavMenuItem";
