+++ src/components/molecules/NavBar/NavBar.style.tsx
@@ -2,19 +2,19 @@ import styled from "styled-components";
 
 import { MainTheme } from "@/theme/theme";
 
-export type StyledNavBarItemContainerProps = {
+export type StyledNavBarContainerProps = {
   theme?: MainTheme;
   gap?: number;
 };
-export const StyledNavBarItemContainer = styled.ol<StyledNavBarItemContainerProps>`
+export const StyledNavBarContainer = styled.ol<StyledNavBarContainerProps>`
   display: flex;
   align-items: center;
   justify-content: start;
   padding: 0 16px;
   width: 100%;
   font-size: 16px;
-  font-family: ${({ theme }: StyledNavBarItemContainerProps) =>
+  font-family: ${({ theme }: StyledNavBarContainerProps) =>
     theme?.font.family.livvic};
-  gap: ${({ gap }: StyledNavBarItemContainerProps) => `${gap}px`};
+  gap: ${({ gap }: StyledNavBarContainerProps) => `${gap}px`};
   list-style-type: none;
 `;
