+++ src/components/molecules/index.ts
@@ -1,3 +1,4 @@
+export * from "./BurgerMenu/BurgerMenu";
 export * from "./DashBoardTag/DashboardTag";
-export * from "./NavBarDesktop/NavBarDesktop";
+export * from "./NavBar/NavBar";
 export * from "./ProfileActivity/ProfileActivity";
